requirements.txt
novus[vbu,speed]
pillow
numpy
typing_extensions
asyncpg
git+https://github.com/Voxel-Fox-Ltd/VoxelFoxCheckout
